backend/requirements.txt
fastapi
uvicorn
httpx
pymongo
python-multipart
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2